db-migration:change id key back to int
