+++ cimis/alembic/versions/d4b31059af3d_.py
@@ -1,0 +1,30 @@
+"""empty message
+
+Revision ID: d4b31059af3d
+Revises: 8e87ac56acc7
+Create Date: 2021-06-17 18:47:27.422238
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import mssql
+
+# revision identifiers, used by Alembic.
+revision = 'd4b31059af3d'
+down_revision = '8e87ac56acc7'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('DailyRaw', sa.Column('Id', sa.Integer(), nullable=False))
+    op.add_column('HourlyRaw', sa.Column('Id', sa.Integer(), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('HourlyRaw', 'Id')
+    op.drop_column('DailyRaw', 'Id')
+    # ### end Alembic commands ###
